--- images/banana1-gr-obj-bkg
@@ -1,4 +1,0 @@
-object:
-(150,50), (150,51), (150,52), (150,53), (150,54), (150,55), (150,56), (150,57), (150,58), (150,59), (150,60), (150,61),
-background:
-(), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), ()
